refactor(ui): split mesh viewport policy
